+++ src/thornfield/caches/memory_cache.py
@@ -1,3 +1,5 @@
+from typing import Any
+
 from .cache import Cache
 from .volatile_value import VolatileValue
 from ..constants import NOT_FOUND
@@ -5,10 +7,10 @@ from ..constants import NOT_FOUND
 
 class MemoryCache(Cache):
     def __init__(self) -> None:
-        super().__init__(serialization_needed=False)
+        super().__init__()
         self._cache = {}
 
-    def get(self, key):
+    def get(self, key: str) -> Any:
         if key not in self._cache:
             return NOT_FOUND
         value = self._cache[key]
@@ -16,7 +18,7 @@ class MemoryCache(Cache):
             return self._from_volatile(value)
         return value
 
-    def set(self, key, value, expiration: int):
+    def set(self, key: str, value: Any, expiration: int) -> None:
         if expiration:
             value = self._to_volatile(value, expiration)
         self._cache[key] = value