+++ src/thornfield/caches/postgresql_cache.py
@@ -1,5 +1,5 @@
 from time import time
-from typing import Optional, Any, Callable
+from typing import AnyStr
 
 from .cache import Cache
 from ..constants import NOT_FOUND
@@ -11,30 +11,28 @@ from ..postgresql_key_value_adapter import (
 
 
 class PostgresqlCache(Cache):
-    def __init__(
-        self,
-        connection_pool: ConnectionPool,
-        table: str,
-        serializer: Optional[Callable[[Any], str]] = None,
-        deserializer: Optional[Callable[[Optional[str]], Any]] = None,
-    ) -> None:
-        super().__init__(serializer=serializer, deserializer=deserializer)
+    def __init__(self, connection_pool: ConnectionPool, table: str) -> None:
+        super().__init__()
         self._adapter = PostgresqlKeyValueAdapter(connection_pool, table)
 
-    def get(self, key):
-        t = round(time() * 1000)
+    def get(self, key: str):
+        t = self._get_curr_time()
         try:
-            value = self._adapter.get(self._serialize(key), t)
+            value = self._adapter.get(key, t)
             if value is None:
                 return NOT_FOUND
-            return self._deserialize(value)
+            return value
         except Exception as e:
             raise CachingError(f"Could not get {key}", exc=e)
 
-    def set(self, key, value, expiration: int):
+    def set(self, key: str, value: AnyStr, expiration: int) -> None:
         if expiration:
-            expiration = round(time() * 1000) + expiration
+            expiration = self._get_curr_time() + expiration
         try:
-            self._adapter.set(self._serialize(key), self._serialize(value), expiration)
+            self._adapter.set(key, value, expiration)
         except Exception as e:
             raise CachingError(f"Could not set {key} as {value}", exc=e)
+
+    @staticmethod
+    def _get_curr_time():
+        return round(time() * 1000)